refactor(git): extract prewarm logic into service_prewarm.go (#328)
